+++ src/aac_datasets/datasets/audiocaps.py
@@ -280,19 +280,16 @@ class AudioCaps(Dataset[Dict[str, Any]]):
     # Magic methods
     def __getitem__(
         self,
-        index_and_column: Any,
+        index: Any,
     ) -> Dict[str, Any]:
         if (
-            isinstance(index_and_column, tuple)
-            and len(index_and_column) == 2
-            and (
-                isinstance(index_and_column[1], (str, Iterable))
-                or index_and_column[1] is None
-            )
+            isinstance(index, tuple)
+            and len(index) == 2
+            and (isinstance(index[1], (str, Iterable)) or index[1] is None)
         ):
-            index, column = index_and_column
+            index, column = index
         else:
-            index, column = index_and_column, None
+            column = None
 
         item = self.at(index, column)
         if self.__transform is not None: